+++ sdk/lib/health/index.ts
@@ -1,3 +1,2 @@
 import "./checkFns"
-
 import "./HealthReceipt"
